fix: Handle downcast_ref in Option case
When accessing a field of an entity, it's possible that the object
associated with the parent value may be of type Option<T> instead of T.
Make sure this is handled correctly, at least in the case of
entity_object_via_relation where is_owner = true. I'm not sure if the
issue can arise in other cases but it might need fixing there too.