Add support for custom subscription fields (#214)
Although Seaography doesn't directly implement subscriptions itself, a
user that wants to implement subscriptions needs a way to have these
included in the schema. The build() function of async_graphql::dynamic's
Schema type requires that if subscriptions are going to be used, the
name of the top-level subscription type be specified. Since this
function is called from within Seaography's Builder::new() function,
there is no other way for users to add subscriptions later.

To allow for this use case, have Builder maintain a vector of
SubscriptionField objects that are supplied by the user, and include
these in a Subscription object that is part of the schema. This object
is *only* added if the vector is non-empty, since empty object types are
not allowed by the GraphQL spec.