Return null for None, don't resolve subfields
This is a follow-up to #209.

When the return type of a custom query or mutation is Option<T>, and at
runtime a particular resolution of that field returns None, stop there
and return null in the response without trying to resolve subfields of
the type.

What was previously happening in this case was that a value of the form
Some(FieldValue::owned_any(x)) was returned, where x was a Rust value of
type Option<T>. We would then try to resolve fields on this value, which
lead to two problems:

1. Whenever an attempt was made to call try_downcast_ref::<T>, this
would fail because the owned object actually has the type Option<T>. My
initial attempt to fix this first tried downcasting to T, and if that
failed, instead try downcasting to Option<T>. This worked in cases where
the value was actually Some, but still resulted in errors when the value
was None.

2. A None value would still have its fields resolved, so for example if
you did a query like staff_by_id(id: 0) { firstName }, it would return
the result:

    "staff_by_id": { "firstName": null }

This is incorrect, because it implies that a staff object with the
specified id actually exists, but has a firstName value of null. In
reality, the object does not exist, and the correct response is:

    "staff_by_id": null

This PR fixes the problem at the source, changing the gql_field_value()
function called on the results of custom queries and mutations to return
Option<FieldValue> instead of FieldValue, and having the code generated
by #[derive(CustomOperation)] return either Some or None depending on
the result of gql_field_value(). With this change, when a None value is
returned, no attempt is made to resolve any of the fields on the
corresponding type, meaning we no longer need to attempt downcasting to
Option<T>, and the problem of downcasting on a None value cannot occur.