Output conversions: Proper error & None handling (#217)
Modify the type signature of FnOutputTypeConversion so that instead of
returning Value, it returns an optional FieldValue. This makes it
possible to provide conversion functions which return None (indicating
the value does not exist and should not have its subfields resolved),
and also to return other kinds of FieldValues, such as those which own a
Rust value (via FieldVale::owned_any).

The need for this came from several cases where we have columns stored
in the database as json, but in our entity definitions have column types
that are custom Rust structs specific to our application, for which we
implement the appropriate traits to convert to and from both
serde_json::Value and sea_orm::Value. In some cases these fields are
nullable, e.g. we have a column which is of type Option<AccountConfig>.

This change allows us to implement the relevant conversion functions so
they can return None if the account config is null, or
Some(FieldValue::owned_any(config)) if the account config is non-null,
after deserializing the AccountConfig struct from json.