Allow GraphQL types to be specified per column
Add extra fields to TypesMapConfig to allow a user of the API to specify
the input and output types given in the schema for specific columns.
When combined with suitable conversion functions, this is particularly
useful for providing type safety for columns that are stored in the
database as JSON, but which have user-defined Rust types in the entity
definition.

Rust types used for these fields need to implement certain traits from
SeaORM in order to actually be used, but that is outside the scope of
this PR. The present change just allows the types to be given in the
schema. This makes it possible (in fact required) to specify subfields
in queries, and to obtain type safety when using suitable client
libraries.

I have some other changes that (at least in my view) make the API easier
to use, but i'll address those in a separate PR as they aren't strictly
necessary for obtaining the functionality described above.