Don't require TypesMapHelper in ctx.data (#221)
The implementation of singular queries relied on the assumption that a
TypesMapHelper object would be available in the data associated with an
async_graphql context. This isn't actually necessary, because it's
possible to create a new TypesMapHelper instance given a 'static
reference to BuilderContext, which is already available to the field
implementation.

This change avoids requiring setup code that creates a schema from
having to register a TypesMapHelper instance with the schema builder. As
far as I can tell, this is the only place in seaography where this was
an issue.