Add options for default + max pagination limit (#223)
* Add options for default + max pagination limit

Add fields to PaginationInputConfig that allow default and maximum
limits to be specified, which affect the way in which PaginatedInput
arguments to queries are handled. Both are optional, and are not set by
default.

Specifying the maximum limit can be used to protect against requests
that pose undue load on the server due to queries that could otherwise
return tens or hundreds of thousands of rows.

Specifying the default limit allows clients to opt-in to pagination,
subject to the maximum limit (if specified). If the maximum limit is set
but the default is not, the maximum limit will be considered the default
in the case where no limit is specified.

If the default or maximum is set, and a client does not specify any
pagination options in a query, the 'page' pagination mode will used,
with a page number of 0.

* Add tests

---------

Co-authored-by: Chris Tsang <chris.2y3@outlook.com>