Tidy up per-column type options
Combine the multiple BTreeMap fields of TypesMapConfig into one field,
column_options. Each entry in this map is a ColumnOptions object
containing the various options that until now were stored in separate
BTreeMaps.

This makes it easier to set multiple options at once. For example, if
you have a specific set of options that you want to apply to multiple
columns, you only only need to create a single ColumnOptions object and
then apply that to several columns. To facilitate this, ColumnOptions
implements the Clone trait, and the conversion function types have been
changed from Box to Arc to allow them to be cloned.

The ColumnOptions struct is declared as #[non_exhaustive] so that new
fields can be added to it in future versions of seaography without
breaking backwards compatibility. This attribute prevents creation of
instances using the syntax ColumnOptions { ... } and instead requires an
instance to be constructed with default() and then modified. The latter
approach will not break with the addition of new fields.

Whether the column_options field should belong in TypesMapConfig or
BuilderContext is with considering. Currently I've made it part of the
former since it's presently only used for type-related options, but it's
possible that other kinds of options might be added in the future, in
which case putting it in BuilderContext might make more sense.

The column_options BTreeMap is indexed by keys of the newly-introduced
EntityColumnId type. This is an opaque identifier for an (entity,
column) pair which can only be created by specifying the Entity type and
Column value from the SeaORM definitions (see example below). The
purpose of using this instead of strings is to enforce a compile-time
guarantee that the columns referenced actually exist. This avoids the
following problems that can easily occur when referencing the names by
strings:

- Misspelt entity or column names

- Capitalization of entity or column names that differs from that
  produced by the type_name and column_name functions in
  EntityObjectConfig.

- Renaming of columns in the SeaORM entity definition and forgetting to
  update the config

Usage example:

    fn owned_column_options<T>(
        input_type: Option<TypeRef>,
        output_type: Option<TypeRef>,
    ) -> ColumnOptions
    where
        T: for<'de> Deserialize<'de> + Send + Sync + 'static,
    {
        let mut options = ColumnOptions::new();
        options.input_conversion = Some(Arc::new(convert_value_from_json));
        options.output_conversion = Some(Arc::new(convert_value_to_owned::<T>));
        options.input_type = input_type;
        options.output_type = output_type;
        options
    }

    /// ...

    context.types.column_options.insert(
        EntityColumnId::of::<accounts::Entity>(&accounts::Column::Config),
        owned_column_options::<AccountConfig>(
            Some(TypeRef::named("AccountConfigInput")),
            Some(TypeRef::named("AccountConfig")),
        ),
    );