Macros for custom types & fields
This PR adds a much richer system of types (and associated macros) for
use in custom fields, as well as entity fields whose type is overridden.

I will write a more detailed description shortly, but the main ones are:

CustomInputType (trait and derive macro): defines gql_input_type_ref and
parse_value. The trait can be derived on structs and enums (the latter
being either graphql enums or graphql unions)

CustomOutputType (trait and derive macro): defines gql_output_type_ref
and gql_field_value. The trait can be derived on structs and enums. The
trait can be derived on structs and enums (the latter being either
graphql enums or graphql unions)

CustomFields (trait and attribute macro): Indended as a replacement for
CustomOperation. Discovers fields on the impl block only, avoiding the
need to list them separately as members of the associated struct. Also
allows access to instances, by way of the self parameter which can
optionally be included in fields.

This commit is just an initial version so you can see the code. Soon I
will add some examples and test cases which show them in action.

We have been using these traits and macros in our own codebase for some
time (just with different names) and all our own integration tests pass
with them. Some of these will form the basis of integration tests that
I'll add to seaography itself.