Initial version of sea-draw example app (#227)
* Initial version of sea-draw example app

This app demonstrates a simple but practical example backend for a web
app. It follows the same patterns for using seaography that we've used
in the real backend for the Rush platform, and exercises various
features that we rely on such as custom types, mutations, and
subscriptions.

Currently this is a work-in-progress; this is just the initial version
and in the coming days I'll add some missing functionality and write up
proper documentation.  Hopefully this will serve not just as a test case
but also a useful example from which future users of seaography can use
as a guide to building their own apps.

The app contains a data model appropriate for a (very simplistic)
collaborative vector drawing app. A drawing contains a set of objects,
each of which has a shape, position, stroke, and fill. Drawings are
grouped into projects. Permissions indicating which accounts have what
level of access to projects (not yet implemented) will be used for
access control.

Some of the things it covers (not all implemented yet):
- Support for multiple database backends (sqlite and postgres)
- Custom create/update/delete mutations that operate on a per-entity
  basis rather than mapping more directly to SQL statements
- Validation on create/update operations
- Soft deletes (setting a deleted_at field on an entity)
- Access control and a permissions mechanism based on accounts &
  projects (accounts identified based on Authorization headers)
- Use of JSONB columns for storing data that in Rust is represented as
  structs and enums, with appropriate trait impls generated by the
  custom type macros.
- Support for custom fields on entities (e.g. Drawing.svg)
- Generation of notification events on changes to entities that are
  transmitted to interested clients via a GraphQL subscription

* Tidy up formatting

* Fix custom fields for shapes

* sea-draw: Permission checks + more tests

* Fix missing root account in sqlite schema

* cargo fmt

* Add sea-draw integration tests

* Split out sea-draw tests

* cleanup unused code

* Fix build

* Re-enable warnings

* Implement permission checks for queries

---------

Co-authored-by: Chris Tsang <chris.2y3@outlook.com>