Add 'not' operator to filters (#231)
Seaography already supports 'and' and 'or' operators on filters. This
commit adds a 'not' operator, which takes a filter object as an argument
and maps this onto the corresponding Condition not() method in SeaORM.

This is particularly useful for using operators that do not have an
inverse version, such as array contains. For example, in one use case,
we want to select all entities whose tags field does _not_ contain a
specific value:

    {
      runs(filters: {
          not: { tags: { array_contains: ["disabled"] } }
      }) {
        nodes {
          id
          name
          status
        }
      }
    }